requirements.txt
torch==2.5.0
torchvision==0.20.0
tqdm
scikit-image
matplotlib
Pillow
numpy